ggjylb-DS.v2.5.3.2.5/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
autofeat==2.1.1
certifi==2025.6.15
charset-normalizer==3.4.2
choreographer==1.0.9
click==8.2.1
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
docker==7.1.0
eli5==0.16.0
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
fastjsonschema==2.21.1
flexcache==0.3
flexparser==0.4
fonttools==4.58.4
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaleido==1.0.0
kiwisolver==1.4.8
llvmlite==0.44.0
logistro==1.1.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.43.1
nbformat==5.10.4
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
orjson==3.10.18
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
Pint==0.24.4
platformdirs==4.3.8
plotly==6.1.2
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
shap==0.48.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
xgboost==3.0.2
